May 10, 2024, 3:37 PM

+++ src/app/categories/page.js
@@ -16,8 +16,23 @@ const CategoriesPage = () => {
             .then(data => setCategories(prev => prev = data))
     }
 
-    async function handleRemove(id) {
+    function preventDelete(category) {
+        if (category === "spicy"
+            || category === "salty"
+            || category === "traditional"
+            || category === "france"
+            || category === "chinese") {
+            return true
+        }
+        else return false
+    }
+
+    async function handleRemove(id, category) {
+        if (preventDelete(category)) {
+            return toast.error("no no, you cant delete my categories")
+        }
         const response = await fetch("/api/categories", {
+
             method: "DELETE",
             body: JSON.stringify({
                 _id: id
@@ -73,16 +88,21 @@ const CategoriesPage = () => {
                 <form onSubmit={handleSubmitCategories} className='flex flex-col md:w-[500px] py-5 gap-y-4'>
                     <label htmlFor="category" className=' flex font-bold  w-full '>New Category</label>
                     <input type="text" className='flex rounded-full shadow-sm py-4 border px-3 font-bold grow uppercase ' value={newCategory} onChange={e => setNewCategory(e.target.value)} />
-                    <input type="submit" className='bg-sekunder hover:bg-slate-200 cursor-pointer' />
+                    <input type="submit" value={"ADD"} className='bg-sekunder hover:bg-slate-200 cursor-pointer' />
                 </form>
                 <div className=' flex  w-full justify-center items-center'>
                     <ul className='flex flex-col gap-4'>
                         {
                             categories && categories?.map((category) => (
                                 <li key={category._id} className='flex justify-between items-center outline-dashed md:w-[500px] w-[300px] pl-5 '>{
-                                    category?.name
-                                }
-                                    <p className='font-extrabold h-full p-4 hover:bg-red-500 cursor-pointer duration-300' onClick={() => handleRemove(category?._id)}>X</p>
+                                    preventDelete(category?.name) ?
+                                        <div className='justify-between w-full flex'>
+                                            {category.name}   < span className='text-slate-400 font-serif'> default</span>
+                                        </div> : category?.name
+                                } <span>
+
+                                    </span>
+                                    <p className='font-extrabold h-full p-4 hover:bg-red-500 cursor-pointer duration-300' onClick={() => handleRemove(category?._id, category?.name)}>X</p>
                                 </li>
                             ))
